localflask/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
gunicorn
lxml